Add a custom hologram for Auto Splitters
We need a custom hologram, as the default hologram just crahes when
trying to upgrade (it does something weird and probably wrong with the
connections).

So we inhibit those changes and also whatever is happening in the
conveyor attachment hologram, which wants to snap to everything,
apparently.

Unfortunately, this custom hologram needs a patch for the original
hologram to make a private member with a list of the old building's
connections available.