Switch to fixed precision for rates
In order to avoid constant uncontrollable round-off errors during rate
calculations, switch all rates to fixed precision with a predetermined
number of fractional digits (3 for now).

The interface to the widget still operates on floats. This change
requires replacing the old, float-based output rates which will already
have ended up in people's savegames with a new property for
integer-based rates. The old property has been deprecated and will
automatically be migrated as part of the version 0 -> version 1
migration of the loaded splitter.