Reworked distribution algorithm
The distribution algorithm now works in two stages: In step 1, all
requirements are collected in an upward direction towards the root of
the network, and in a second step the actual allocated rates are
distributed in top-down fashion.

The algorithm seems to work in auto mode, testing for fixed rates still
to come.