Make Close button work on widget
Something is really weird about the window dark widget, it does not
allow us to access the on close button pressed event. So this justs
inherits a new blueprint from the window dark widget and re-exports the
event. Then it's just a matter of forwarding the event to the escape
pressed event of the interact widget.